build: add code generation script
Add a code generation script and build steps in the Dockerfile.dapper
This allows the GRPC code generation to happen inside the dapper build
environment, which makes it independent of the developers host
environment to a greated degree. It is inspired by the equivalent
mechanism of the longhorn-engine and helps to avoid headaches when
updating GRPC code due to protobuf version discrepancies and similar
problems.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>